reorder generation eval

+++ src/main/java/genetics/fitness/Fitness.java
@@ -2,5 +2,5 @@ package genetics.fitness;
 
 public abstract class Fitness {
     public Fitness(){}
-    public abstract int fit(float[] output, float[] expected);
+    public abstract float fit(float[] output, float[] expected);
 }
